+++ pyproject.toml
@@ -22,13 +22,15 @@ authenticated_a2a_client = "datarobot_genai.dragent.plugins.auth_a2a_client"
 okta_cross_app_access = "datarobot_genai.dragent.plugins.okta_a2a_auth"
 per_user_tool_calling_agent = "datarobot_genai.dragent.plugins.per_user_tool_calling_agent"
 streaming_memory_agent = "datarobot_genai.dragent.plugins.streaming_memory_agent"
-datarobot_llm_providers = "datarobot_genai.nat.datarobot_llm_providers"
-datarobot_llm_clients = "datarobot_genai.nat.datarobot_llm_clients"
 datarobot_mcp_client = "datarobot_genai.nat.datarobot_mcp_client"
 datarobot_auth_provider = "datarobot_genai.nat.datarobot_auth_provider"
 datarobot_mem0_memory = "datarobot_genai.nat.datarobot_mem0_memory"
 datarobot_moderation_middleware = "datarobot_genai.nat.datarobot_moderation_middleware"
 datarobot_otelcollector = "datarobot_genai.dragent.plugins.datarobot_otelcollector"
+llm_providers = "datarobot_genai.dragent.plugins.llm_providers"
+llm_clients = "datarobot_genai.dragent.plugins.llm_clients"
+crewai_llm_clients = "datarobot_genai.crewai.llm_clients"
+llamaindex_llm_clients = "datarobot_genai.llama_index.llm_clients"
 
 [project.entry-points.'nat.front_ends']
 dragent = "datarobot_genai.dragent.frontends.register"
