+++ tests/dragent/plugins/test_langchain_llm_clients.py
@@ -15,14 +15,8 @@
 import os
 from unittest.mock import patch
 
-import datarobot_genai.nat.datarobot_llm_clients  # noqa: F401
 import pytest
 from crewai import LLM
-from datarobot_genai.nat.datarobot_llm_providers import DataRobotLitellmConfig
-from datarobot_genai.nat.datarobot_llm_providers import DataRobotLLMComponentModelConfig
-from datarobot_genai.nat.datarobot_llm_providers import DataRobotLLMDeploymentModelConfig
-from datarobot_genai.nat.datarobot_llm_providers import DataRobotLLMGatewayModelConfig
-from datarobot_genai.nat.datarobot_llm_providers import DataRobotNIMModelConfig
 from langchain_core.language_models import BaseChatModel
 from langchain_litellm import ChatLiteLLM
 from llama_index.llms.litellm import LiteLLM
@@ -30,6 +24,11 @@ from nat.builder.framework_enum import LLMFrameworkEnum
 from nat.builder.workflow_builder import WorkflowBuilder
 
 from datarobot_genai.core.config import LLMType
+from datarobot_genai.dragent.plugins.llm_providers import DataRobotLitellmConfig
+from datarobot_genai.dragent.plugins.llm_providers import DataRobotLLMComponentModelConfig
+from datarobot_genai.dragent.plugins.llm_providers import DataRobotLLMDeploymentModelConfig
+from datarobot_genai.dragent.plugins.llm_providers import DataRobotLLMGatewayModelConfig
+from datarobot_genai.dragent.plugins.llm_providers import DataRobotNIMModelConfig
 
 
 async def test_datarobot_llm_gateway_langchain():
@@ -40,26 +39,6 @@ async def test_datarobot_llm_gateway_langchain():
         await builder.add_llm("datarobot_llm", llm_config)
         llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LANGCHAIN)
         assert isinstance(llm, BaseChatModel)
-
-
-async def test_datarobot_llm_gateway_crewai():
-    llm_config = DataRobotLLMGatewayModelConfig(
-        model_name="azure/gpt-4o-2024-11-20", temperature=0.0, api_key="some_token"
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-
-
-async def test_datarobot_llm_gateway_llamaindex():
-    llm_config = DataRobotLLMGatewayModelConfig(
-        model_name="azure/gpt-4o-2024-11-20", temperature=0.0, api_key="some_token"
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
 
 
 async def test_datarobot_llm_deployment_langchain():
@@ -86,66 +65,6 @@ async def test_datarobot_llm_deployment_langchain_with_identity_token():
         assert llm.extra_headers == {"X-DataRobot-Identity-Token": "identity-token-123"}
 
 
-async def test_datarobot_llm_deployment_crewai():
-    llm_config = DataRobotLLMDeploymentModelConfig(
-        llm_deployment_id="123", temperature=0.0, api_key="some_token"
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-
-
-async def test_datarobot_llm_deployment_crewai_with_identity_token():
-    llm_config = DataRobotLLMDeploymentModelConfig(
-        llm_deployment_id="123",
-        temperature=0.0,
-        api_key="some_token",
-        headers={"X-DataRobot-Identity-Token": "identity-token-123"},
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-        assert llm.additional_params == {
-            "max_retries": 10,
-            "extra_headers": {"X-DataRobot-Identity-Token": "identity-token-123"},
-            "stream_options": {"include_usage": True},
-        }
-
-
-async def test_datarobot_llm_deployment_llamaindex():
-    llm_config = DataRobotLLMDeploymentModelConfig(
-        llm_deployment_id="123", temperature=0.0, api_key="some_token"
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
-
-
-async def test_datarobot_llm_deployment_llamaindex_with_identity_token():
-    # Pin endpoint so test is independent of .env (e.g. DATAROBOT_ENDPOINT)
-    with patch.dict(
-        os.environ, {"DATAROBOT_ENDPOINT": "https://app.datarobot.com/api/v2"}, clear=False
-    ):
-        llm_config = DataRobotLLMDeploymentModelConfig(
-            llm_deployment_id="123",
-            temperature=0.0,
-            api_key="some_token",
-            headers={"X-DataRobot-Identity-Token": "identity-token-123"},
-        )
-        async with WorkflowBuilder() as builder:
-            await builder.add_llm("datarobot_llm", llm_config)
-            llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
-        assert llm.additional_kwargs == {
-            "api_base": "https://app.datarobot.com/api/v2/deployments/123/chat/completions",
-            "api_key": "some_token",
-            "extra_headers": {"X-DataRobot-Identity-Token": "identity-token-123"},
-        }
-
-
 async def test_datarobot_nim_langchain():
     llm_config = DataRobotNIMModelConfig(
         nim_deployment_id="123",
@@ -157,32 +76,6 @@ async def test_datarobot_nim_langchain():
         await builder.add_llm("datarobot_llm", llm_config)
         llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LANGCHAIN)
         assert isinstance(llm, BaseChatModel)
-
-
-async def test_datarobot_nim_crewai():
-    llm_config = DataRobotNIMModelConfig(
-        nim_deployment_id="123",
-        model_name="azure/gpt-4o-2024-11-20",
-        temperature=0.0,
-        api_key="some_token",
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-
-
-async def test_datarobot_nim_llamaindex():
-    llm_config = DataRobotNIMModelConfig(
-        nim_deployment_id="123",
-        model_name="azure/gpt-4o-2024-11-20",
-        temperature=0.0,
-        api_key="some_token",
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("datarobot_llm", llm_config)
-        llm = await builder.get_llm("datarobot_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
 
 
 async def test_datarobot_llm_component_langchain_use_gateway():
@@ -403,29 +296,6 @@ async def test_datarobot_llm_component_llamaindex(
             assert llm.additional_kwargs == {}
 
 
-# --- DataRobotLitellmConfig (external litellm) clients ---
-
-
-async def test_litellm_crewai():
-    llm_config = DataRobotLitellmConfig(model_name="openai/gpt-4o", api_key="test-key")
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("ext_llm", llm_config)
-        llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-        assert llm.model == "openai/gpt-4o"
-
-
-async def test_litellm_crewai_with_temperature():
-    llm_config = DataRobotLitellmConfig(
-        model_name="openai/gpt-4o", api_key="test-key", temperature=0.3
-    )
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("ext_llm", llm_config)
-        llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-        assert llm.temperature == 0.3
-
-
 async def test_litellm_langchain():
     llm_config = DataRobotLitellmConfig(model_name="openai/gpt-4o", api_key="test-key")
     async with WorkflowBuilder() as builder:
@@ -446,15 +316,6 @@ async def test_litellm_langchain_with_temperature():
         assert llm.temperature == 0.7
 
 
-async def test_litellm_llamaindex():
-    llm_config = DataRobotLitellmConfig(model_name="openai/gpt-4o", api_key="test-key")
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("ext_llm", llm_config)
-        llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
-        assert llm.model == "openai/gpt-4o"
-
-
 async def test_litellm_llamaindex_with_temperature():
     llm_config = DataRobotLitellmConfig(
         model_name="openai/gpt-4o", api_key="test-key", temperature=0.9
@@ -466,17 +327,6 @@ async def test_litellm_llamaindex_with_temperature():
         assert llm.temperature == 0.9
 
 
-async def test_litellm_crewai_only_passes_set_fields():
-    """Ensure only explicitly set fields are forwarded (exclude_unset=True)."""
-    llm_config = DataRobotLitellmConfig(model_name="anthropic/claude-3", api_key="key-abc")
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("ext_llm", llm_config)
-        llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.CREWAI)
-        assert isinstance(llm, LLM)
-        assert llm.model == "anthropic/claude-3"
-        assert llm.api_key == "key-abc"
-
-
 async def test_litellm_langchain_only_passes_set_fields():
     """Ensure only explicitly set fields are forwarded (exclude_unset=True)."""
     llm_config = DataRobotLitellmConfig(model_name="anthropic/claude-3", api_key="key-abc")
@@ -484,14 +334,4 @@ async def test_litellm_langchain_only_passes_set_fields():
         await builder.add_llm("ext_llm", llm_config)
         llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.LANGCHAIN)
         assert isinstance(llm, BaseChatModel)
-        assert llm.model == "anthropic/claude-3"
-
-
-async def test_litellm_llamaindex_only_passes_set_fields():
-    """Ensure only explicitly set fields are forwarded (exclude_unset=True)."""
-    llm_config = DataRobotLitellmConfig(model_name="anthropic/claude-3", api_key="key-abc")
-    async with WorkflowBuilder() as builder:
-        await builder.add_llm("ext_llm", llm_config)
-        llm = await builder.get_llm("ext_llm", wrapper_type=LLMFrameworkEnum.LLAMA_INDEX)
-        assert isinstance(llm, LiteLLM)
         assert llm.model == "anthropic/claude-3"
